Revert "se ve en la ruta /map no lo he enrrutado para no meterme con los componentes de rafa "
