+++ client/src/App.jsx
@@ -11,7 +11,7 @@ import {
   Dashboard,
   MenuForm
 } from './components/components.js';
-import LocalsDatabasic from './components/Locales/LocalsDatabasic';
+import Mapdata from './components/Map/Map';
 import { Routes, Route } from 'react-router-dom';
 import { useDispatch, useSelector } from 'react-redux';
 import { checkUser } from './redux/actions/actions';
@@ -36,7 +36,7 @@ function App () {
         <Route path='/answers' element={<Answers />} />
         <Route path='/about' element={<About />} />
         <Route path='/dataTreatment' element={<DataTreatment />} />
-        <Route path='/map' element={<LocalsDatabasic />} />
+        <Route path='/map' element={<Mapdata />} />
         <Route path='/dashboard' element={<Dashboard />} />
         <Route path='/menu' element={<MenuForm />} />
         <Route exact path='/' element={<Landing />} />
