--- client/src/components/Map/SearchMap/SearchMap.js
@@ -1,19 +1,0 @@
-const apiUrl = 'https://nominatim.openstreetmap.org/search';
-const cityName = 'Ciudad de México';
-const SearchMap = (data) => {
-
-  return fetch(`${apiUrl}?q=${data}&format=json`)
-    .then(response => response.json())
-    .then(data => {
-      const latitude = data[0].lat;
-      const longitude = data[0].lon;
-    //   console.log(`Las coordenadas de ${cityName} son: (${latitude}, ${longitude})`);
-   return[latitude,longitude]
-    })
-    .catch(error => {
-      console.log(error);
-    });
-
-
-}
-export default SearchMap
