--- client/src/redux/actions/local.js
@@ -1,57 +1,0 @@
-export const SUCCESS = 'SUCCESS';
-export const DETAIL = 'DETAIL';
-
-export const createLocal = (inputs) => {
-    return async (dispatch) => {
-      try {
-        const response = await axios.post('/locals', {
-          email: inputs.email,
-          images: inputs.images,
-          location: inputs.location.location,
-          lat:inputs.location.lat,
-          lng:inputs.location.lng,
-          name: inputs.name,
-          // phone: inputs.phone,
-          schedule: inputs.schedule,
-          specialty: inputs.specialty
-          // characteristics: chekinputs
-        });
-        if (response.status === 201) {
-          dispatch({
-            type: SUCCESS,
-            payload: response.data.success
-          });
-          setTimeout(() => {
-            dispatch({
-              type: SUCCESS_RESET
-            });
-          }, 3000);
-        }
-      } catch (error) {
-        console.log(error);
-        dispatch({
-          type: ERROR,
-          payload: error.message
-        });
-        // set error state to null after 3 seconds
-        setTimeout(() => {
-          dispatch({
-            type: ERROR_RESET
-          });
-        }, 3000);
-      }
-    };
-  };
-export const DetailLocal = (id) => {
-  return async (dispatch) => {
-    try {
-      const datos = await axios.get(`/locals/${id}`);
-      dispatch({
-        type: DETAIL,
-        payload: datos.data
-      });
-    } catch (error) {
-      console.log(error.message);
-    }
-  };
-};
