+++ client/src/redux/reducer/reducer.js
@@ -1,7 +1,6 @@
 
-import { ORDER, RESET, CREATE_USER, HOMEPAGE,  ERROR, SUCCESS_RESET, ERROR_RESET, CHECKUSER, RESETUSER } from '../actions/actions';
-import { SEARCH_BY_QUERY } from '../actions/cards';
-import { DETAIL, SUCCESS } from '../actions/local';
+import { ORDER, RESET, SEARCH_BY_QUERY, DETAIL, CREATE_USER, HOMEPAGE, SUCCESS, ERROR, SUCCESS_RESET, ERROR_RESET, CHECKUSER, RESETUSER } from '../actions/actions';
+
 const initialState = {
   cards: [],
   reset: [],
